Introduce light and app client modes.

+++ src/app_client.rs
@@ -2,7 +2,7 @@ use std::sync::mpsc::Receiver;
 
 use crate::types::ClientMsg;
 
-pub async fn run(app_id: u16, block_receive: Receiver<ClientMsg>) {
+pub async fn run(app_id: u32, block_receive: Receiver<ClientMsg>) {
 	log::info!("Starting for app {}...", app_id);
 
 	for block in block_receive {
